+++ src/components/Post/PostItem/PostItem.jsx
@@ -85,44 +85,11 @@ const BtnMore = styled.button`
   top: 7px;
   right: 0;
 `;
-export default function PostItem({ modalOpen }) {
+export default function PostItem({ modalOpen, postInfo, authorInfo }) {
   const navigate = useNavigate();
   function moveDetail() {
     navigate("/detailpost");
   }
-  const [postInfo, setPostInfo] = useState([]);
-  const [authorInfo, setAuthorInfo] = useState([]);
-
-  useEffect(() => {
-    getUserInfo();
-  }, []);
-  const getUserInfo = async () => {
-    const token = localStorage.getItem("token");
-    const accountname = localStorage.getItem("accountname");
-    console.log(token);
-    const res = await axios.get(
-      `https://api.mandarin.weniv.co.kr/post/${accountname}/userpost`,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-type": "application/json",
-        },
-      },
-    );
-    // console.log("여기");
-    console.log("여기", res.data.post[0].author);
-    // const { username, image } = res.data.post.author;
-    // setUserInfo({
-    //   username,
-    //   image,
-    // });
-    const posts = res.data.post;
-    const authors = res.data.post[0].author;
-    setPostInfo(posts);
-    setAuthorInfo(authors);
-  };
-  console.log(postInfo);
-  console.log(authorInfo);
 
   return (
     <>
